Possible pose detection algorithm
This works pretty well when there's an arrow in the middle of a full size sheet of paper (print included arrow in middle of paper, don't cut it out).  However, it lets through a lot of false positives, as seen when using an arrow on a small card size piece of paper or when the camera is just looking at me. It's pretty speedy though, so maybe we can look into applying a mask and getting the color average like https://stackoverflow.com/questions/54316588/get-the-average-color-inside-a-contour-with-open-cv or something